Doctrine injection
