+++ app/Model/Database/Entity/SerialEntity.php
@@ -5,83 +5,55 @@ namespace App\Model\Database\Entity;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Entity
- * @ORM\Table(name="serial")
- */
+#[ORM\Entity]
+#[ORM\Table(name: "serial")]
 final class SerialEntity
 {
 
-	/**
-	 * @ORM\Id
-	 * @ORM\GeneratedValue(strategy="AUTO")
-	 * @ORM\Column(type="integer", nullable=false)
-	 */
+	#[ORM\Id]
+	#[ORM\GeneratedValue(strategy: "AUTO")]
+	#[ORM\Column(type: "integer", nullable: false)]
 	protected int $id;
 
-	/**
-	 * @ORM\Column(type="string", nullable=false)
-	 */
+	#[ORM\Column(type: "string", nullable: false)]
 	protected string $slug;
 
-	/**
-	 * @ORM\Column(type="string")
-	 */
+	#[ORM\Column(type: "string")]
 	protected string $name;
 
-	/**
-	 * @ORM\Column(type="string", nullable=false)
-	 */
+	#[ORM\Column(type: "string", nullable: false)]
 	protected string $year;
 
-	/**
-	 * @ORM\Column(type="smallint", nullable=true)
-	 */
+	#[ORM\Column(type: "smallint", nullable: true)]
 	protected ?int $rating;
 
-	/**
-	 * @ORM\Column(type="text", nullable=false)
-	 */
+	#[ORM\Column(type: "text", nullable: false)]
 	protected string $teaser;
 
-	/**
-	 * @ORM\Column(type="text", nullable=false)
-	 */
+	#[ORM\Column(type: "text", nullable: false)]
 	protected string $description;
 
-	/**
-	 * @ORM\Column(type="text", nullable=false)
-	 */
+	#[ORM\Column(type: "text", nullable: false)]
 	protected string $imagePoster;
 
-	/**
-	 * @ORM\Column(type="text", nullable=false)
-	 */
+	#[ORM\Column(type: "text", nullable: false)]
 	protected string $imageBanner;
 
-	/**
-	 * @ORM\Column(type="text", nullable=false)
-	 */
+	#[ORM\Column(type: "text", nullable: false)]
 	protected string $trailer;
 
-	/**
-	 * @var Collection<int, UserEntity>
-	 * @ORM\ManyToMany(targetEntity="UserEntity", mappedBy="likeSerial")
-	 * @ORM\JoinTable(name="serial_like")
-	 */
+	/** @var Collection<int, UserEntity> */
+	#[ORM\ManyToMany(targetEntity: UserEntity::class, mappedBy: "likeSerial")]
+	#[ORM\JoinTable(name: "serial_like")]
 	protected Collection $likeUser;
 
-	/**
-	 * @var Collection<int, UserEntity>
-	 * @ORM\ManyToMany(targetEntity="UserEntity", mappedBy="laterSerial")
-	 * @ORM\JoinTable(name="serial_later")
-	 */
+	/** @var Collection<int, UserEntity> */
+	#[ORM\ManyToMany(targetEntity: UserEntity::class, mappedBy: "laterSerial")]
+	#[ORM\JoinTable(name: "serial_later")]
 	protected Collection $laterUser;
 
-	/**
-	 * @var Collection<int, SerialLastEntity>
-	 * @ORM\OneToMany(targetEntity="SerialLastEntity", mappedBy="serial")
-	 */
+	/** @var Collection<int, SerialLastEntity> */
+	#[ORM\OneToMany(mappedBy: "serial", targetEntity: SerialLastEntity::class)]
 	protected Collection $serialLast;
