+++ app/Model/Database/Fixture/ArticleFixture.php
@@ -3,7 +3,6 @@
 namespace App\Model\Database\Fixture;
 
 use App\Model\Database\Entity\ArticleEntity;
-use App\Model\Facade\Admin\Content\Month\MonthFacade;
 use Doctrine\Persistence\ObjectManager;
 use Nette\Neon\Neon;
 use Nette\Utils\Strings;
@@ -11,13 +10,6 @@ use Nettrine\Fixtures\ContainerAwareInterface;
 
 final class ArticleFixture extends AbstractFixture implements ContainerAwareInterface
 {
-
-	public function getMonthFacade(): MonthFacade
-	{
-		return $this->getContainer()->getByType(MonthFacade::class);
-	}
-
-
 	public function load(ObjectManager $manager): void
 	{
 		$articles = Neon::decodeFile(__DIR__ . "/content/article.neon");
@@ -30,7 +22,6 @@ final class ArticleFixture extends AbstractFixture implements ContainerAwareInte
 			$newArticle->setDescription($article['description']);
 			$newArticle->setImage($article['image']);
 			$newArticle->setCreatedAt($timeNow);
-			$newArticle->setCreatedAtMonth($this->getMonthFacade()->getMonth($timeNow));
 			$manager->persist($newArticle);
 		}
 		$manager->flush();
